refactor: Revert article fragment to old version
